Initial cut of the command framework with reply-to included and a sample JMS implementation.
I've not tested any of this yet - I just wanted to create a prototype to see how it fit together & if the Command & Dispatcher paradigm would work well.

Its looking OK. As we drill down we'll see if this is a leaky abstraction - I think its gonna be OK.

To get up to speed try the MasterController and SlaveController in the jms package. Bad name Controller - might be better to call it Service or Agent or something.