Properties-based hook for including system-tests into a JUnit run.
maven test:single -Dtestcase=org.sysunit.SumoSystemTestSuite

will look for /sysunit-tests.properties

sysunit-tests.properties includes a line for each foo.jelly to be
run as part of the system-test.

ie:

	/foo/bar/bazTest.jelly
	/foo/bar/cheeseTest.jelly

If you'd like to selectively disable some, add a value to the RHS

	/foo/bar/bazTest.jelly
	/foo/bar/cheeseTest.jelly = off

The test is considered to be enabled if the RHS is:

	blank
	"yes"
	"true"
	"on"
	"enabled"

Otherwise, the test will be considered disabled.

Will shortly be adding a maven goal to create the sysunit-tests file
on-the-fly (optionally not clobbering a custom one).

It -should- also look for the property org.sysunit.sumo.test-list
for the name of the file to use, as either a classpath resource
or a java.io.File, but maven isn't passing -Dorg.sysunit.sumo.test-list
along for some odd reason.  Possibly the dash, maybe.  Will figure that
out, so that you can maintain subsets of system-tests.

For now, just edit src/test/sysunit-tests.properties in your project
to add/remove/enable/disable tests.